projects/synth_tabular_classification/requirements.txt
# Core ML (CPU-only for testing)
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
joblib>=1.3.0

# Experiment tracking
mlflow>=2.10.0

# Data versioning
dvc[gs]>=3.30.0

# API server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Linting
ruff>=0.1.0
mypy>=1.7.0
